pyproject.toml
[project]
name = "retail-sales-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.8",
    "dbt-core>=1.11.7",
    "dbt-snowflake>=1.11.3",
    "pandas>=2.1.2,<3.0.0",
    "plotly>=6.6.0",
    "python-dotenv>=1.2.2",
    "snowflake-connector-python[pandas]>=4.3.0",
    "streamlit>=1.55.0",
]
